Add Find Command Feature
