+++ src/main/java/app/monobot/MonoBotInputParser.java
@@ -49,10 +49,22 @@ public class MonoBotInputParser extends InputParserEventSource {
             case "unmark":
                 this.ProcessUnmarkInput(split);
                 break;
+            case "find":
+                this.ProcessFindInput(split);
+                break;
             default:
                 this.InvokeErrorEvent("Unknown Command! :o");
                 break;
         }
+    }
+
+    private void ProcessFindInput(String[] split) {
+        if (split.length < 2) {
+            this.InvokeErrorEvent("Find search keyword is missing! :o");
+            return;
+        }
+        String keyword = split[1];
+        this.bot.PrintFindTaskList(keyword);
     }
 
     private void ProcessTodoInput(String[] split) {
